switch coastlines from openstreetmapdata.com to osmdata.openstreetmap.de
openstreetmapdata.com is not longer active. We take the water-polygons from osmdata.openstreetmap.de.
For Zoom>8 there is no difference, for lower zoom level there is now only one shape file for all zoom levels, but there is no visible difference after rendering.
These are the same coastlines as used for osm carto, so we have a chance they are updated for long time.